MAJOR UPDATE: All existing bugs resolved!
The "#include" for the main header file was not on the correct line for all .cpp files.

The Initialize_Program class needed to be loaded last in the header file.

I finalized what libraries and .dll files to use for the MySQL Connector and have copied all .dll files into both the Debug and Release folders.

Project 'Configuration Properties' are now properly set.

Had to specify "std::string" in the main header file to avoid conflicts with the <mysqlx/xdevapi.h> file's string object.