UPDATE: CHECKPOINT ACHIEVED!!! Base code functionality + debugging complete! Currently, 3 UIs are fully functional: start, login, and main menu. I've reworked how to handle incorrect user entries to avoid any unwanted bugs/crashes.
Made most variables in main() file GLOBAL for that file. This made implementing functions much easier and practical. It also allowed me to clean up my main.cpp a lot!

Fixed a bug in the closeConn() method: the variable "conn" was not being set to false. So, it was constantly set to "true".

In "Menus.cpp", I reworked the SCRN methods. Before, each method would set "prevMenu" to the current menu before changing the current menu. This created an issue if the menu/UI were to be displayed twice in a row. An endless loop could be processed for the BASE screen. This has been resolved.