UPDATE: Major changes and functionality added! Multiple new UIs created and setup to branch from Main Menu. More sql functionality added: connect to local database, create Table object, retrieve data from queries and display that info on the UI.
WARNING: None of this has been tested due to a single error  being thrown (Code: C2665). It is coming form "memory" and I currently have no idea how to resolve this. I've tested my code in a sandbox and it worked just fine. Will have to investigate further to figure out what is causing this issue.