UPDATE: insertModMovieData() now has error-handling and a definitive process for adding a new movie to the database.
This function is treated as a looping process. It must be completed before the user can return to the previous menu.

Much has been added to this function, but I'm putting a pause on this until it's migrated to the new class "Admin_Actions()". Then, I need to rework how it handles errors. No reason for user to re-enter all prompts if one is entered incorrectly.
Also, need to allow user to press [0] to exit out of this process at any time in any prompt.

TO-DO -->
1. Create a Process on the DB to assist in processing the user inputs.
2. Migrate new functions from main() to a new class. New class shall handle specific dialogs generated from the Admin Actions menu.
3. Add feature that displays what menu/UI the user is currently in.
4. Add functionality for ADD/UPDATE/DELETE options in the Display Table UIs.
5. CHANGE: All tables shall be manually editable via Display Table UI. Not just some of them. HOWEVER... warn the user that their actions may cause problems in the DB if certain data is modified/deleted.
6. Add new feature in Display Table UI: list all associated table names to the current table being display and list the type of association. In other words, inform the user how their actions may affect other tables in the DB.