UPDATE:
Program now checks to see if an actor exists in the database prior to inserting a new one to avoid duplication errors.

Fixed a bug where the user selected "[1] Add/Remove Movie" in the Admin Actions menu and it started the "Add New Movie" process isntead of displaying the next UI.

Removed any data pertaining to ADD/UPDATE/DELETE in the Display_Table UI.

TO-DO:

Transfer prior work from laptop that I did on the trip. That work contains updates for reworking how directors are handled during the "add movie" process.

Apply same functionality from setCast() function to setDir() function. Must verify that we are not entering duplicate people into the system.