UPDATE: Long Change Log!!!
Minor changes to comments.
Minor changes to displayed text (corrections to mispelling and incorrect references).
Minor changes to menu options that are available during the "add new movie" process.
Minor change to Error Handling logic in setNumCast() and setNumDir().
Removed [tblIndexes] from setGenre() function. Was not being used.

Issue with setRating() -> If user entered shorthand (i.e., nc17 or pg13 instead of nc-17 or pg-13), program recorded shorthand version. Issue resolved - program now records full rating.

Changes to addGenre() function -> Issue with potitioning of try-catch blocks and while loop. Split into 2 try-catch blocks. While loop is now outside of try-catch which is inside loop now.

Issue discovered and resolved with [setCast()] -> When user adds new actor to the database, this change is not updated on the terminal screen. User is not able to select newly added actor. This was resolved by reworking the [setCast()], [insertMovieData()], and [modifyNewMovie()] functions to repeatedly refresh tabledata and call the [setCast()] function until all necessary conditions have been met.
Function [setCast()] now takes in a Schema object to make things easier.
To assist in resolving this issue, a new function [addNewCast()] was created!
Var [newCastMembers] removed since it was no longer being used. Because of this, var [oldCastMembers] was renamed to its original name [movieCastMembers]. Consequently, the var [oldDirectors] was renamed to its original name [movieDirectors] and the var [newDirectors] was deleted.

Function [addGenre()] name changed to [addNewGenre()] to better indicate its purpose.

Confirmed previous changes: when modifying user data during the "add new movie" process, all available options are fully accessible.

BUGS:

001: During "add new movie" process, when user is adding a new genre to the database, program does NOT search for duplicate entries before sending new genre to the DB.

TO-DO:

Code has changed so much that my code changes on my laptop are mostly obsolete. I will apply the same functionality from [setCast()] to [setDir()] in a future git push.

When selecting genres to link to the new movie, add an option to CLEAR the user selection and start over (in case user makes a mistake).

Add a title to the header indicating that the user in in the process of "adding a new movie" to the database.