UPDATE:
Copied the functionality from [setCast()] to [setDir()]. Changes from laptop were out of date and no longer usable.

New function [addNewDir()] was created to assist with the new functionality of [setDir()].

Since passing the Schema object from one function to the next seemed to be working really well, I applied this concept to the functions [setTitle()] and [setGenre()].
During this process, some changes were made to [addNewGenre()] so that it aligned similarly with the functions [addNewCast()] and [addNewDir()].

One of the error messages shared between [addNewCast()] and [addNewDir()] was in charge of indicating when a duplicate was detected (when user attempted to add a new person to the DB). The name was displayed kindof weirdly, so I made it look better.

BUGS:

001: When adding new genres to the database, program does NOT check for duplicate entries first.

TO-DO:

Apply similar changes to [validationSCRN()] function which were done to [setTitle()] and [setGenre()]. Incorporate Schema object to reduce on code.

Provide search function for actors + directors: should be available in both "Display Tables" menu/UI and during "add new movie" process.

When selecting genres to link to the new movie, add an option to CLEAR the user selection and start over (in case user makes a mistake).