UPDATE:
I was able to quickly resolve the two bugs encountered by the previous Git push. The rework done in [main.cpp] is complete and fully functional now. While some things were moved around and some code was changed, the functionality of the code was not changed. In fact, I discovered that some of my error handling in the main() function was not working as intended. That has been resolved.

I deleted my leftover 'test' code that I wrote in the last Git push which helped me verify that the functions in [main.cpp] were properly handling errors. Some of this 'test' code extended to [Menus.cpp] and was removed from there too.

The next step is to work on the primary TO-DO list one item at a time to slowly improve the program's capabilities and services.