UPDATE: Tested program on a different machine and confirmed that the special RGB/ANSI color codes do work! The issue was most likely due to the fact that I was developing on a Windows Server PC and not a standard version of Windows.
I added the rest of the Dracula themed colors to the [Menus.cpp] file.

NOTE: One bug exists -> when the program is initially launched, the background of the text is the correct color, but the rest of the terminal is black. As soon as the user triggers any type of reponse from the program, the entire background is updated with the correct color.